+++ src/app/locales/en/components/space-automation/atm-workflow-executions-table.js
@@ -7,6 +7,7 @@ export default {
     scheduledAt: 'Scheduled at',
     startedAt: 'Started at',
     finishedAt: 'Finished at',
+    suspendedAt: 'Suspended at',
     status: 'Status',
     actions: '',
   },
@@ -14,6 +15,7 @@ export default {
     waiting: 'There are no waiting workflow executions',
     ongoing: 'There are no ongoing workflow executions',
     ended: 'There are no ended workflow executions',
+    suspended: 'There are no suspended workflow executions',
   },
   workflowRow,
 };