+++ src/app/services/workflow-actions.js
@@ -1,7 +1,6 @@
 /**
  * Provides workflows manipulation functions ready to use for GUI.
  *
- * @module services/workflow-actions
  * @author Michał Borzęcki
  * @copyright (C) 2021 ACK CYFRONET AGH
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
@@ -9,6 +8,7 @@
 
 import Service from '@ember/service';
 import CancelAtmWorkflowExecutionAction from 'oneprovider-gui/utils/workflow-actions/cancel-atm-workflow-execution-action';
+import PauseResumeAtmWorkflowExecutionAction from 'oneprovider-gui/utils/workflow-actions/pause-resume-atm-workflow-execution-action';
 
 export default Service.extend({
   /**
@@ -22,5 +22,18 @@ export default Service.extend({
    */
   createCancelAtmWorkflowExecutionAction(context) {
     return CancelAtmWorkflowExecutionAction.create({ ownerSource: this, context });
+  },
+
+  /**
+   * @param {Object} context context specification:
+   *   ```
+   *   {
+   *     atmWorkflowExecution: Models.AtmWorkflowExecution|Models.AtmWorkflowExecutionSummary,
+   *   }
+   *   ```
+   * @returns {Utils.WorkflowActions.PauseResumeAtmWorkflowExecutionAction}
+   */
+  createPauseResumeAtmWorkflowExecutionAction(context) {
+    return PauseResumeAtmWorkflowExecutionAction.create({ ownerSource: this, context });
   },
 });