+++ CHANGELOG.md
@@ -6,6 +6,8 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-9767 Added support for new automation workflow execution states
+* VFS-9824 Removed "OnedataFS credentials" automation type
 * VFS-9339 Added showing rest api for file
 * VFS-9614 Added archivisation audit log
 * VFS-9902 Fixed GUI crash when a long path to file is shortened